requirements.txt
tkcalendar~=1.6.1
cups~=0.0.6
barcode~=1.0.4
pillow~=11.1.0
python-dateutil~=2.9.0.post0
requests~=2.32.3
selenium~=4.29.0
customtkinter~=5.2.2
isbnlib~=3.10.14
scikit-learn~=1.6.1
Jinja2~=3.1.6